src/test/java/com/leetcode/strings/easy/GoatLatin_824Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class GoatLatin_824Test {

  private final GoatLatin_824 solution = new GoatLatin_824();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments("I speak Goat Latin", "Imaa peaksmaaa oatGmaaaa atinLmaaaaa"),
          arguments(
              "The quick brown fox jumped over the lazy dog",
              "heTmaa uickqmaaa rownbmaaaa oxfmaaaaa umpedjmaaaaaa overmaaaaaaa hetmaaaaaaaa azylmaaaaaaaaa ogdmaaaaaaaaaa"),
          arguments(
              "Each word consists of lowercase and uppercase letters only",
              "Eachmaa ordwmaaa onsistscmaaaa ofmaaaaa owercaselmaaaaaa andmaaaaaaa uppercasemaaaaaaaa etterslmaaaaaaaaa onlymaaaaaaaaaa"));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void toGoatLatin(String str, String expectedResult) {
    String actualResult = solution.toGoatLatin(str);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
